requirements.txt
autogen==0.3.1
openai==1.44.0
chromadb<=0.5.0
markdownify==0.13.1
ipython==8.18.1
pypdf==5.0.1
psycopg-binary==3.2.3
psycopg-pool==3.2.3
sentence_transformers==3.3.0
python-dotenv==1.0.1
geopy==2.4.1